backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastembed
fastapi==0.114.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
chromadb==0.5.5
sentence-transformers==2.6.1
torch==2.3.1+cpu
python-multipart==0.0.9
pypdf==4.2.0
python-docx==1.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.3
lxml==5.2.1
requests==2.32.3
numpy>=1.24.0
PyPDF2==3.0.1


